Extract cache related methods to new class
